Seperated out Transform comp to its own header
