+++ Sol/Source/Sol/Scene/Components/TransformComp.h
@@ -1,0 +1,42 @@
+#pragma once
+#include <glm/glm.hpp>
+#include <glm/gtx/quaternion.hpp>
+#include "GalaxyDraw/GalaxyMacros.h"
+#include "GalaxyDraw/SceneCamera.h"
+#include <GalaxyDraw/Interfaces/Texture.h>
+#include <GalaxyDraw/Interfaces/Model.h>
+#include "Sol/SolDefines.h"
+
+namespace Sol
+{
+	struct TransformComp
+	{
+	public:
+
+		TransformComp() = default;
+		TransformComp(const TransformComp&) = default;
+		TransformComp(const glm::vec3 position) : Position(position) { }
+		TransformComp(const glm::vec3 position, const glm::vec3 rotation) : Position(position), Rotation(rotation) { }
+		TransformComp(const glm::vec3 position, const glm::vec3 rotation, const glm::vec3 scale) : Position(position), Rotation(rotation), Scale(scale) { }
+		~TransformComp() = default;
+
+		glm::mat4 GetTransformMatrix() const
+		{
+			glm::mat4 rotation = glm::toMat4(glm::quat(Rotation));
+
+			return glm::translate(glm::mat4(1.0f), Position)
+				* rotation
+				* glm::scale(glm::mat4(1.0f), Scale);
+		}
+
+		operator const glm::mat4() const {
+			return GetTransformMatrix();
+		}
+
+
+		glm::vec3 Position = { 0.0f, 0.0f, 0.0f };
+		glm::vec3 Rotation = { 0.0f, 0.0f, 0.0f };
+		glm::vec3 Scale = { 1.0f, 1.0f, 1.0f };
+
+	};
+}
